requirements-dev.txt
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
Pillow==5.2.0
pycodestyle==2.4.0